LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/reporting/access-reporting/property-opener-set.https-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: SecurityError: Blocked a frame with origin "http://localhost:8800" from accessing a cross-origin frame. Protocols, domains, and ports must match.

Harness Error (TIMEOUT), message = null

TIMEOUT same-origin > w => w.opener = "" Test timed out
NOTRUN cross-origin > w => w.opener = ""
